% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stats.R
\name{add_stats}
\alias{add_stats}
\title{Add statistical output}
\usage{
add_stats(
  list,
  stats,
  title = NULL,
  fn = NULL,
  labels = NULL,
  notes = NULL,
  ...
)
}
\arguments{
\item{list}{A list to which the statistical output will be added.}

\item{stats}{The statistical output that will be added to the list.}

\item{title}{A short string describing the statistical output.}

\item{fn}{An optional function used to format the statistical output.}

\item{labels}{The column used to label data frame rows (or a character vector
).}

\item{notes}{A string containing additional information.}

\item{...}{Additional arguments passed on to \code{fn()}.}
}
\value{
A list containing the statistical output appended to the existing object.
}
\description{
Add statistical output
}
