% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_stats}
\alias{format_stats}
\title{Title}
\usage{
format_stats(stats, labels = NULL, ...)
}
\arguments{
\item{stats}{The statistical output that will formatted.}

\item{labels}{The column used to label data frame rows (or a character vector
with length equal to the number of rows).}

\item{...}{Additional arguments passed on to \code{tidy()} and \code{glance()} from the
\code{broom} package.}
}
\value{
A list containing the formatted statistical output.
}
\description{
Title
}
