% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{write_stats}
\alias{write_stats}
\title{Title}
\usage{
write_stats(list, file)
}
\arguments{
\item{list}{A list of statistical output generated using \code{add_stats()}.}

\item{file}{A character string naming a file to write the statistical output.}
}
\description{
Title
}
